#7 - Fixes tilting bug in heatmap
There was a bug in graphology-svg that was tilting the graph based on
the signs of its barycenter. This issue is solved in v0.1.2 which has
been updated. At this point, I also added a node reducer to always tilt
the graph on the y axis, to get a SVG orientation matching the sigma and
heatmap orientation.